Fixed characteristic miscounts
