+++ app/client/src/components/shared/TribalMapList.js
@@ -47,7 +47,6 @@ import {
 } from 'contexts/locationSearch';
 import { LayersProvider, useLayers } from 'contexts/Layers';
 import {
-  useCharacteristicsByGroupContext,
   useServicesContext,
   useStateNationalUsesContext,
 } from 'contexts/LookupFiles';
@@ -970,7 +969,6 @@ function MonitoringTab({
 
   const { monitoringLocations } = useMonitoringLocations();
   const { monitoringLocationsLayer } = useLayers();
-  const characteristicsByGroup = useCharacteristicsByGroupContext();
 
   // sort the monitoring locations
   const [monitoringLocationsSortedBy, setMonitoringLocationsSortedBy] =
@@ -1114,7 +1112,6 @@ function MonitoringTab({
           >
             <div css={accordionContentStyles}>
               <WaterbodyInfo
-                characteristicsByGroup={characteristicsByGroup}
                 type="Past Water Conditions"
                 feature={feature}
                 services={services}
