+++ app/client/src/utils/utils.ts
@@ -1,54 +1,5 @@
 // types
-import type { CharacteristicGroupMappings } from 'config/characteristicGroupMappings';
 import type { KeyboardEvent, MouseEvent } from 'react';
-import type {
-  AnnualStationData,
-  MonitoringLocationAttributes,
-  MonitoringPeriodOfRecordData,
-} from 'types';
-
-interface PeriodOfRecordDatum {
-  Provider: string;
-  MonitoringLocationIdentifier: string;
-  YearSummarized: number;
-  CharacteristicType: string;
-  CharacteristicName: string;
-  ActivityCount: number;
-  ResultCount: number;
-  OrganizationIdentifier: string;
-}
-
-// Add the stations' historical data to the `dataByYear` property,
-export function addAnnualData(
-  monitoringLocations: MonitoringLocationAttributes[],
-  annualData: MonitoringPeriodOfRecordData['sites'],
-) {
-  return monitoringLocations.map((location) => {
-    const id = location.uniqueId;
-    if (id in annualData) {
-      return {
-        ...location,
-        dataByYear: annualData[id],
-        // Tally characteristic counts
-        totalsByCharacteristic: Object.values(annualData[id]).reduce(
-          (totals, yearData) => {
-            Object.entries(yearData.totalsByCharacteristic).forEach(
-              ([charc, count]) => {
-                if (count <= 0) return;
-                if (charc in totals) totals[charc] += count;
-                else totals[charc] = count;
-              },
-            );
-            return totals;
-          },
-          {} as { [characteristic: string]: number },
-        ),
-      };
-    } else {
-      return location;
-    }
-  });
-}
 
 // utility function to split up an array into chunks of a designated length
 export function chunkArray(array: any, chunkLength: number): Array<Array<any>> {
@@ -104,76 +55,6 @@ export function formatNumber(number: number, digits: number = 0) {
     minimumFractionDigits: digits,
     maximumFractionDigits: digits,
   });
-}
-
-// Helper function to get label of characteristic group
-function getCharcLabel(
-  charcGroup: string,
-  labelMappings: CharacteristicGroupMappings,
-) {
-  for (let mapping of labelMappings) {
-    if (mapping.groupNames.includes(charcGroup)) return mapping.label;
-  }
-  return 'Other';
-}
-
-function incrementTotals(
-  station: AnnualStationData,
-  record: PeriodOfRecordDatum,
-  mappings: CharacteristicGroupMappings,
-) {
-  station.totalMeasurements += record.ResultCount;
-  station.totalSamples += record.ActivityCount;
-  if (!(record.CharacteristicName in station.totalsByCharacteristic)) {
-    station.totalsByCharacteristic[record.CharacteristicName] = 0;
-  }
-  station.totalsByCharacteristic[record.CharacteristicName] +=
-    record.ResultCount;
-  if (!(record.CharacteristicType in station.totalsByGroup)) {
-    station.totalsByGroup[record.CharacteristicType] = 0;
-  }
-  station.totalsByGroup[record.CharacteristicType] += record.ResultCount;
-  const label = getCharcLabel(record.CharacteristicType, mappings);
-  if (!(label in station.totalsByLabel)) {
-    station.totalsByLabel[label] = 0;
-  }
-  station.totalsByLabel[label] += record.ResultCount;
-}
-
-export function structurePeriodOfRecordData(
-  records: PeriodOfRecordDatum[],
-  mappings: CharacteristicGroupMappings,
-) {
-  let minYear = Infinity;
-  let maxYear = 0;
-  const results: MonitoringPeriodOfRecordData['sites'] = {};
-  records.forEach((record) => {
-    const year = record.YearSummarized;
-    if (year < minYear) minYear = year;
-    if (year > maxYear) maxYear = year;
-    const id =
-      `${record.MonitoringLocationIdentifier}-` +
-      `${record.Provider}-` +
-      `${record.OrganizationIdentifier}`;
-    // Structure the relevant fields of the
-    // returned data, keyed by id
-    if (!(id in results)) {
-      results[id] = {};
-    }
-    if (!(year in results[id])) {
-      results[id][year] = {
-        uniqueId: id,
-        totalMeasurements: 0,
-        totalSamples: 0,
-        totalsByCharacteristic: {},
-        totalsByGroup: {},
-        totalsByLabel: {},
-      };
-    }
-    incrementTotals(results[id][year], record, mappings);
-  });
-  if (minYear > maxYear) minYear = 0;
-  return { minYear, maxYear, sites: results };
 }
 
 export function isAbort(error: unknown) {
